Adding option to download and remove old files

+++ scripts/run.py
@@ -12,18 +12,22 @@ import scripts.datasets_download as dl
 np.random.seed(1)
 
 
-def run_training_iterations(subject, database, epochs, batch_size, subsamp_rate, early_patience, moves, n_iterations=3):
+def run_training_iterations(subject, database, epochs, batch_size,
+                            subsamp_rate, early_patience, moves,
+                            n_iterations, must_down=True, must_remove=False):
     iteration_list = []
     training_time_list = []
     test_accuracy_list = []
     peak_mem_list = []
-    dl.download_subject(subject, int(database[-1]))
+    if must_down:
+        dl.download_subject(subject, int(database[-1]))
     dcl = DeepConvLstm(subject, database, epochs=epochs, learn_rate=0.001, verbose=0,
                        timeback_reach=0.25, batch_size=batch_size, subsamp_rate=subsamp_rate,
                        moves=moves)
     for n_iter in range(n_iterations):
-        os.system("rm -r results/logs/db{}/*".format(database[-1]))
-        os.system("rm -r results/weights/db{}/*".format(database[-1]))
+        if must_remove:
+            os.system("rm -r results/logs/db{}/*".format(database[-1]))
+            os.system("rm -r results/weights/db{}/*".format(database[-1]))
         training_time, test_accuracy, profile_summary, conf_matrix = dcl.run_training(early_patience=early_patience)
         iteration_list.append({
             "training_time": training_time,
